Создал компонет фаворитной кнопки оффера и помистил их в карточки
