+++ src/components/card-offer-near/card-offer-near.tsx
@@ -2,6 +2,7 @@ import { Link} from 'react-router-dom';
 import { Offer } from '../../mock/offers/offer-mocks';
 import {AppRoute} from '../../const';
 import { useEffect } from 'react';
+import FavoriteButton from '../favorite-button/favorite-button';
 
 type CardOfferProps = {
   offer: Offer;
@@ -36,19 +37,9 @@ function CardOfferNear ({offer}: CardOfferProps): JSX.Element {
             <b className="place-card__price-value">€{offer.price}</b>
             <span className="place-card__price-text">/&nbsp;night</span>
           </div>
-          <button
-            className={`place-card__bookmark-button ${offer.isFavorite ? 'place-card__bookmark-button--active' : ''} button`}
-            type="button"
-          >
-            <svg
-              className="place-card__bookmark-icon"
-              width={18}
-              height={19}
-            >
-              <use xlinkHref="#icon-bookmark" />
-            </svg>
-            <span className="visually-hidden">In bookmarks</span>
-          </button>
+
+          <FavoriteButton offer={offer}/>
+
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
