+++ src/components/card-offers/card.tsx
@@ -2,7 +2,7 @@ import {Link} from 'react-router-dom';
 import {useState} from 'react';
 import type {Offer} from '../../mock/offers/offer-mocks';
 import {AppRoute} from '../../const';
-
+import FavoriteButton from '../favorite-button/favorite-button';
 
 type CardPagesProps = {
   offer: Offer;
@@ -13,16 +13,9 @@ type CardPagesProps = {
 
 function CardOffer ({offer, handleIdOffer, onLeaveMouseOffer}: CardPagesProps) : JSX.Element{
 
-  const [isFavoriteCard, setIsFavoriteCard] = useState(offer.isFavorite);
   const [cardState, setCardState] = useState({
     offerId: offer.id
   });
-
-  const onFavoriteButton = (): void => {
-    const updatedIsFavorite = !isFavoriteCard;
-    setIsFavoriteCard(updatedIsFavorite);
-    offer.isFavorite = updatedIsFavorite;
-  };
 
   function onGetIdCard () {
     setCardState({
@@ -60,12 +53,9 @@ function CardOffer ({offer, handleIdOffer, onLeaveMouseOffer}: CardPagesProps) :
             <b className="place-card__price-value">&euro;{offer.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button onClick={onFavoriteButton} className={`place-card__bookmark-button ${offer.isFavorite ? 'place-card__bookmark-button--active' : ''} button`} type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">In bookmarks</span>
-          </button>
+
+          <FavoriteButton offer={offer}/>
+
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
