+++ src/components/favorite-card-offers/favorite-card-offers.tsx
@@ -1,4 +1,5 @@
 import type { Offer } from '../../mock/offers/offer-mocks';
+import FavoriteButton from '../favorite-button/favorite-button';
 
 type FavoriteOfferProps = {
   offer: Offer;
@@ -22,12 +23,9 @@ function FavoriteCardOffer ({offer}: FavoriteOfferProps): JSX.Element {
             <b className="place-card__price-value">&euro;{offer.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className={`place-card__bookmark-button ${offer.isFavorite ? 'place-card__bookmark-button--active' : ''} button`} type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">In bookmarks</span>
-          </button>
+
+          <FavoriteButton offer={offer}/>
+
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
